Recibo como muestras fallidas una lista de enteros con los id de las muestras

+++ src/main/java/com/ttps/laboratorio/dto/request/UrlDTO.java
@@ -19,6 +19,6 @@ public class UrlDTO {
 	@NotBlank(message = "Url is required.")
 	private String url;
 
-	private List<Integer> failedSamples = new ArrayList<>();
+	private List<Long> failedSamples = new ArrayList<>();
 
 }
